socket_io_flask/requirements.txt
eventlet
python-socketio
kombu
redis
socketIO-client===0.7.2
gunicorn
flask
Flask-Script

